broke Arena down into seperate parts
