+++ frontend/src/components/arena/ArenaView.css
@@ -149,6 +149,13 @@
   pointer-events: auto;
 }
 
+.fighter-desc {
+  margin-top: -15px;
+  font-style: italic;
+  font-size: .7em;
+  font-weight: bold;
+}
+
 .introduction {
   margin-top: 25px;
   font-weight: bold;
@@ -172,210 +179,8 @@
   text-transform: capitalize;
 }
 
-.commentator-container {
-  display: inline;
-}
-
 .next-match {
   font-style: italic;
-}
-
-.commentator-img {
-  display: inline;
-  max-height: 160px;
-  width: auto;
-  border: 4px solid #000;
-  margin-bottom: 5px;
-}
-
-.commentator-img.talking {
-  animation: jimTalkBounce 0.4s ease-in-out;
-}
-
-@keyframes slideUpFadeIn {
-  to {
-    transform: translateY(0);
-  }
-}
-
-@keyframes jimTalkBounce {
-
-  0%,
-  100% {
-    transform: scaleY(1) translateY(0);
-  }
-
-  25% {
-    transform: scaleY(1.05) translateY(-5px);
-  }
-
-  50% {
-    transform: scaleY(0.95) translateY(2px);
-  }
-
-  75% {
-    transform: scaleY(1.02) translateY(-2px);
-  }
-}
-
-.betting-module {
-  background: #ffffff;
-  border: 2px solid #444;
-  padding: 3px;
-  margin: 10px auto 30px auto;
-  max-width: 600px;
-  color: rgb(0, 0, 0);
-}
-
-.pool-display h3 {
-  margin: 0 0 15px 0;
-  text-align: center;
-  font-size: 1.5em;
-  letter-spacing: 1px;
-}
-
-.betting-controls {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-.bet-input-row {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 15px;
-  font-weight: bold;
-}
-
-.bet-input-row input {
-  padding: 8px;
-  font-size: 1.1em;
-  border: 1px solid #666;
-  width: 100px;
-  text-align: center;
-}
-
-.wallet {
-  color: #2ecc71;
-}
-
-.bet-buttons {
-  display: flex;
-  justify-content: space-between;
-  gap: 20px;
-}
-
-
-.bet-btn {
-  flex: 1;
-  padding: 6px;
-  border: none;
-  font-size: .8em;
-  font-weight: bold;
-  cursor: pointer;
-  color: white;
-}
-
-.bet-btn:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-  transform: none;
-}
-
-.bet-btn:active:not(:disabled) {
-  transform: scale(0.95);
-}
-
-.p1-bet {
-  background-color: #834d4d;
-  border: 2px solid black;
-}
-
-.p2-bet {
-  background-color: #94c6ff;
-  color: black;
-  border: 2px solid black;
-}
-
-.bet-btn small {
-  display: block;
-  font-weight: normal;
-  font-size: 0.8em;
-  opacity: 0.9;
-  margin-top: 5px;
-}
-
-.login-prompt {
-  text-align: center;
-  color: #aaa;
-  font-style: italic;
-  margin: 0;
-}
-
-.confetti-container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  pointer-events: none;
-  z-index: 9999;
-  overflow: hidden;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.payout-text {
-  position: absolute;
-  font-size: 6em;
-  color: #2ecc71;
-  animation: payoutPopUp 3.5s ease-out forwards;
-  z-index: 10000;
-  font-weight: bolder;
-}
-
-.lost-text {
-  position: absolute;
-  font-size: 6em;
-  color: #ff1212;
-  animation: payoutPopUp 3.5s ease-out forwards;
-  z-index: 10000;
-  font-weight: bolder;
-}
-
-.cash-confetti {
-  position: absolute;
-  top: -150px;
-  animation: moneyFall 3s linear forwards;
-  filter: drop-shadow(0px 5px 5px rgba(0, 0, 0, 0.5));
-}
-
-@keyframes payoutPopUp {
-  0% {
-    opacity: 0;
-  }
-
-  10% {
-    opacity: 1;
-  }
-
-  100% {
-    opacity: 0;
-  }
-}
-
-@keyframes moneyFall {
-  0% {
-    transform: translateY(0vh) rotate(0deg);
-    opacity: 1;
-  }
-
-  100% {
-    transform: translateY(110vh) rotate(720deg);
-    opacity: 0;
-  }
 }
 
 .current-action {
