+++ frontend/src/components/arena/Betting.css
@@ -1,0 +1,102 @@
+.betting-module {
+  background: #ffffff;
+  border: 2px solid #444;
+  padding: 3px;
+  margin: 10px auto 30px auto;
+  max-width: 600px;
+  color: rgb(0, 0, 0);
+}
+
+.pool-display h3 {
+  margin: 0 0 15px 0;
+  text-align: center;
+  font-size: 1.5em;
+  letter-spacing: 1px;
+}
+
+.betting-controls {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+}
+
+.bet-input-row {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 15px;
+  font-weight: bold;
+}
+
+.bet-input-row input {
+  padding: 8px;
+  font-size: 1.1em;
+  border: 1px solid #666;
+  width: 100px;
+  text-align: center;
+}
+
+.wallet {
+  color: #2ecc71;
+}
+
+.bet-buttons {
+  display: flex;
+  justify-content: space-between;
+  gap: 20px;
+}
+
+.bet-btn {
+  flex: 1;
+  padding: 6px;
+  border: none;
+  font-size: .8em;
+  font-weight: bold;
+  cursor: pointer;
+  color: white;
+}
+
+.bet-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+  transform: none;
+}
+
+.bet-btn:active:not(:disabled) {
+  transform: scale(0.95);
+}
+
+.p1-bet {
+  background-color: #834d4d;
+  border: 2px solid black;
+}
+
+.p2-bet {
+  background-color: #94c6ff;
+  color: black;
+  border: 2px solid black;
+}
+
+.bet-btn small {
+  display: block;
+  font-weight: normal;
+  font-size: 0.8em;
+  opacity: 0.9;
+  margin-top: 5px;
+}
+
+.login-prompt {
+  text-align: center;
+  color: #aaa;
+  font-style: italic;
+  margin: 0;
+}
+
+.fade-in-down {
+  animation: fadeInDown 0.4s ease-out forwards;
+}
+
+@keyframes fadeInDown {
+  0% { opacity: 0; transform: translateY(-20px); }
+  100% { opacity: 1; transform: translateY(0); }
+}
